% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspgldrmFit.R
\name{bspgldrmFit}
\alias{bspgldrmFit}
\title{Main MCMC function}
\usage{
bspgldrmFit(
  X,
  y,
  linkfun,
  linkinv,
  mu.eta,
  mb,
  sb,
  dir_pr_parm,
  bspgldrmControl,
  thetaControl
)
}
\description{
This function is called by the main \code{bspgldrm} function.
}
\keyword{internal}
