% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{dir_parm}
\alias{dir_parm}
\title{Title: function for finding dirichlet proposal parameter}
\usage{
dir_parm(y, tht, btht, dir_pr_parm, ind_mt)
}
\arguments{
\item{y}{(vector) response variable values \{y_i, i=1(1)n\}}

\item{tht}{(vector) derived parameter, \{theta_i = function(beta, f0, x_i), i=1(1)n\}}

\item{btht}{(vector)  normalizing constant b(theta) = \{b(theta_i), i=1(1)n\}}

\item{dir_pr_parm}{(vector) dirichlet prior parameter}

\item{ind_mt}{indicator matrix with (i,j)th element: 1(y_i == s_j)}
}
\value{
(vector) dirichlet proposal parameter
}
\description{
Title: function for finding dirichlet proposal parameter
}
\keyword{internal}
