% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{tht_sol}
\alias{tht_sol}
\title{Title: function for finding theta, btheta & bprime2}
\usage{
tht_sol(spt, f0, mu, thtst)
}
\arguments{
\item{spt}{(vector) support, \{s_j, j = 1(1)l\}, of response variable y}

\item{f0}{(vector) \{f0(y = s_j), j=1(1)l\}: free parameter, f0(.) centering density}

\item{mu}{(vector) \{E(y_i | x_i), i=1(1)n\}}

\item{thtst}{(vector) initial value in theta solving iterative algorithm}
}
\value{
(list) containing:
\enumerate{
\item bpr2 (vector) \{second derivative of b(theta_i), i=1(1)n\}
\item tht (vector) derived parameter theta = \{theta_i = function(beta, f0, x_i), i=1(1)n\}
\item btht (vector)  normalizing constant b(theta) = \{b(theta_i), i=1(1)n\}
}
}
\description{
Title: function for finding theta, btheta & bprime2
}
\keyword{internal}
